Change viewport to be pixels instead of ratio.
The previous 'viewport is a ratio' was helpful
but it meant window parameters had to be passed
around to a lot of functions. Now viewports are
just given an rect in pixels.
If the window is resized it is up to the application
to resize the viewport appropriately.
Updated the viewport ray casting to use the pixel size
and convert to ratio.
Also fixed a circular import between viewport -> window
that made the Cocos2d example fail. Some functions were
moved back into viewport from window.