Fix issues in world_transform.
Use getters and setters instead of direct access
to variables.
Remove setter checks for same value.
The way python works for +=, -=, etc means that
we must not check if the value hasn't changed.
Python calls the getter, modifies the value then
passes it back via the setter in these cases.