Rewrite shader to support more advanced features.
The Shader class is now ShaderProgram.
The Shader class now represents a single shader part.
Ie, vertex shader, fragment shader. These are combined
into a ShaderProgram.
Multiple Shaders of the same type can be attached to a
single shader program.
The seperation of Shader and ShaderPrograms also means
we can store collections of pre-compiled Shaders which
are attached to programs and linked. This should improve
performance.