Updates to vertex_buffer.
Swap offset calculation to be simpler and use positive
checks instead of negatives ('if x' instead of 'if not x').
Add checks for the buffer being bound.
Fix check for data not checking against 'None' and causing
numpy errors.
Fix typo in nbyte allocation causing nbytes to become the
'byte' python type.
Fix set_data not honoring the offset value.